+++ CHANGELOG.md
@@ -11,3 +11,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.2]
 
 - Fixing minor issues (package.json, README.md).
+
+## [0.0.3]
+
+- Added first version of auto-completion
+  - Limited to first level properties of the schema (properties of definitions are not used)
+  - Limited to property keys
+  - Examples are indicated
+  - Types are indicated, the $ref is also indicated
+  - Contrainsts are not indicated